samples/code_test_project/requirements.txt
# Python dependencies for CodeRaptor test project

# Core dependencies
pyyaml>=6.0
sqlite3  # Usually included with Python

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional dependencies for extended functionality
requests>=2.28.0
numpy>=1.21.0
pandas>=1.5.0

# Web framework (if building API)
flask>=2.2.0
flask-cors>=3.0.0
gunicorn>=20.1.0
